Upgrade dbt to 1.2.0 (#9)
* Upgrade dbt to 1.2.0

Weirdly this isn't working at the moment -- `poetry install` seems to refuse to recognize that `dbt-core` 1.2.0 exists. I'm having issues in another project upgrading dbt, so possibly there's some issue with the release metadata?

It does not seem to be due to https://github.com/python-poetry/poetry/issues/1930#issuecomment-653906544 in this case.